requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-apispec==2.2.3
aiosignal==1.4.0
apispec==3.3.2
attrs==26.1.0
cffi==2.0.0
cryptography==46.0.7
frozenlist==1.8.0
idna==3.13
Jinja2==3.1.6
libnacl==2.1.0
MarkupSafe==3.0.3
marshmallow==4.3.0
multidict==6.7.1
packaging==26.1
propcache==0.4.1
pyasn1==0.6.3
pycparser==3.0
pyipv8==3.2.0
PyNaCl==1.6.2
pyOpenSSL==26.0.0
python-dotenv==1.2.2
typing_extensions==4.15.0
webargs==5.5.3
yarl==1.23.0
